+++ src/test/lib/until_test.ts
@@ -17,26 +17,53 @@
 
 import {until} from '../../lib/until.js';
 import {html, render} from '../../lit-html.js';
+import {Deferred} from './deferred.js';
 
 const assert = chai.assert;
 
 suite('until', () => {
 
+  let container: HTMLDivElement;
+  let deferred: Deferred<string>;
+
+  setup(() => {
+    container = document.createElement('div');
+    deferred = new Deferred<string>();
+  });
+
   test('displays defaultContent immediately', () => {
-    const container = document.createElement('div');
-    let resolve: (v: any) => void;
-    const promise = new Promise((res, _) => {
-      resolve = res;
-    });
     render(
-        html`<div>${until(promise, html`<span>loading...</span>`)}</div>`,
+        html
+        `<div>${until(deferred.promise, html`<span>loading...</span>`)}</div>`,
         container);
     assert.equal(container.innerHTML, '<div><!----><span>loading...</span><!----></div>');
-    resolve!('foo');
-    return promise.then(() => new Promise((r) => setTimeout(() => r())))
-        .then(() => {
-          assert.equal(container.innerHTML, '<div><!---->foo<!----></div>');
-        });
+    deferred.resolve('foo');
+            return deferred.promise
+                .then(() => new Promise((r) => setTimeout(() => r())))
+                .then(() => {
+                  assert.equal(container.innerHTML, '<div><!---->foo<!----></div>');
+                });
+  });
+
+  test('renders new Promise over existing Promise', () => {
+    const t = (v: any) =>
+        html`<div>${until(v, html`<span>loading...</span>`)}</div>`;
+    render(t(deferred.promise), container);
+    assert.equal(container.innerHTML, '<div><!----><span>loading...</span><!----></div>');
+
+    const deferred2 = new Deferred<string>();
+    render(t(deferred2.promise), container);
+    assert.equal(container.innerHTML, '<div><!----><span>loading...</span><!----></div>');
+
+    deferred2.resolve('bar');
+    return deferred2.promise.then(() => {
+      assert.equal(container.innerHTML, '<div><!---->bar<!----></div>');
+
+      deferred.resolve('foo');
+      return deferred.promise.then(() => {
+        assert.equal(container.innerHTML, '<div><!---->bar<!----></div>');
+      });
+    });
   });
 
 });